+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Upcoming
 
-Nothing here yet.
+### Removed
+
+- Removed `console.log` from `App.svelte`.
 
 ## 1.0.5 - 2024-04-29